Update models accessibility
Re-structure models
Catch 401 status code and throw unauthenticated
Update RequestCompletion<T> = (Result<T, NimbleLoyaltyError>) -> Void
Allow the user to over authenticate
