+++ Sources/Data/NetworkAPI/Models/Public/APIRewardList.swift
